+++ custom_components/teamtracker/const.py
@@ -3,13 +3,7 @@ from homeassistant.const import Platform
 from datetime import timedelta
 
 # API
-URL_HEAD = "http://site.api.espn.com/apis/site/v2/sports/"
-URL_TAIL = "/scoreboard"
 API_LIMIT = 50
-USER_AGENT = (
-    "Mozilla/5.0 (Macintosh; Intel Mac OS X 11_6) AppleWebKit/605.1.15 (KHTML, like "
-    "Gecko) Version/15.0 Safari/605.1.15"
-)
 
 # Config
 CONF_API_LANGUAGE = "api_language"
