+++ custom_components/teamtracker/utils.py
@@ -1,5 +1,6 @@
 """ Miscellaneous Utilities """
 import logging
+import re
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -41,3 +42,16 @@ def has_team(data, target_team_id):
                 if competitor.get("team", {}).get("id") == target_team_id:
                     return True
     return False
+
+
+def season_slug_to_name(slug: str) -> str:
+    """Convert a season slug like '2025-26-english-premier-league' to 'English Premier League'."""
+    if not slug:
+        return ""
+    body = re.sub(r"^\d{4}(-\d{2})?-", "", slug)
+    if body == slug:
+        return ""
+    def _fmt_word(w):
+        # Uppercase abbreviations (no vowels, e.g. "mls", "nfl"); title-case real words
+        return w.upper() if w.isalpha() and not re.search(r"[aeiou]", w, re.I) else w.title()
+    return " ".join(_fmt_word(w) for w in body.split("-"))
